#  nrf5x_common: gpio: fix port 1 functionality 