@PeterKietzmann 2018-10-15
@SemjonKerner can you have a look at this and test it?