@aabadie 2018-10-16
I tested this PR with #9833 and it indeed fixes the problem I reported there with the on-board button (which uses port 1).