@aabadie 2018-10-17
I enabled Murdock because I think this PR is OK. But it would be great to also have @haukepetersen's opinion before merging it.