@haukepetersen 2018-10-17
Nice catch!

Could somebody please also quickly run the GPIO benchmark (`bench` shell command in the `tests/periph_gpio` test application) with this PR and maybe a `nrf52dk`, so we can see if this PR does actually not influence the runtime performance for 1-port nrf CPUs?! If it does change something, we probably want to use a slightly different structure for this PR (e.g. something macro based...).