@bergzand 2018-10-17
Tested the `tests/periph_gpio` on an nrf52dk:
current master (917548f37ce268cf03a044c3e1bf2af8cd15e5c9):
```
> bench 0 1
2018-10-17 15:40:53,436 - INFO #  bench 0 1
2018-10-17 15:40:53,437 - INFO # 
2018-10-17 15:40:53,440 - INFO # GPIO driver run-time performance benchmark
2018-10-17 15:40:53,441 - INFO # 
2018-10-17 15:40:53,560 - INFO #                  nop loop:    109376us  ---   0.109us per call  ---    9142773 calls per sec
2018-10-17 15:40:53,865 - INFO #                  gpio_set:    296876us  ---   0.296us per call  ---    3368409 calls per sec
2018-10-17 15:40:54,122 - INFO #                gpio_clear:    250001us  ---   0.250us per call  ---    3999984 calls per sec
2018-10-17 15:40:54,425 - INFO #               gpio_toggle:    296877us  ---   0.296us per call  ---    3368398 calls per sec
2018-10-17 15:40:54,856 - INFO #                 gpio_read:    421876us  ---   0.421us per call  ---    2370364 calls per sec
2018-10-17 15:40:55,131 - INFO #                gpio_write:    265626us  ---   0.265us per call  ---    3764691 calls per sec
2018-10-17 15:40:55,133 - INFO # 
2018-10-17 15:40:55,136 - INFO #  --- DONE ---
```

With this PR merged on top:
```
> bench 0 1
2018-10-17 15:42:06,996 - INFO #  bench 0 1
2018-10-17 15:42:06,997 - INFO # 
2018-10-17 15:42:07,001 - INFO # GPIO driver run-time performance benchmark
2018-10-17 15:42:07,001 - INFO # 
2018-10-17 15:42:07,118 - INFO #                  nop loop:    109376us  ---   0.109us per call  ---    9142773 calls per sec
2018-10-17 15:42:07,422 - INFO #                  gpio_set:    296876us  ---   0.296us per call  ---    3368409 calls per sec
2018-10-17 15:42:07,681 - INFO #                gpio_clear:    250001us  ---   0.250us per call  ---    3999984 calls per sec
2018-10-17 15:42:07,988 - INFO #               gpio_toggle:    296876us  ---   0.296us per call  ---    3368409 calls per sec
2018-10-17 15:42:08,415 - INFO #                 gpio_read:    421876us  ---   0.421us per call  ---    2370364 calls per sec
2018-10-17 15:42:08,689 - INFO #                gpio_write:    265626us  ---   0.265us per call  ---    3764691 calls per sec
2018-10-17 15:42:08,690 - INFO # 
2018-10-17 15:42:08,692 - INFO #  --- DONE ---
```

Also confirmed with cosy that the gpio.o object files both result in 328B flash