@haukepetersen 2018-10-17
awesome, thanks for verifying this, I guess this PR is ready to go then, right?!

@aabadie I hope you are ok that I set the `3-testing` label in your name...