@haukepetersen 2018-10-17
all green -> go

@silkeh thanks for the fix, I guess that one was on me :-)